Split datastore.import into multiple commands
import.{vmdk,ovf,ova}
